+++ src/models/forms/GoogleAuthenticatorForm.php
@@ -52,6 +52,7 @@ class GoogleAuthenticatorForm extends Model
     public function delete(): false|int
     {
         if ($this->validate()) {
+            Yii::$app->getSession()->set('google_2fa_secret', null);
             $this->user->google_2fa_secret = null;
             return $this->user->update();
         }