+++ CHANGELOG.md
@@ -4,8 +4,7 @@
   methods `getSearchInput`, and `getSearchKeywords` to the new `GridView::$search` property
 - Changed `GridView::isSortedByPosition()` to `GridView::isSortable()`
 - Removed `Html::buttonList()`, use `Html::buttons()` instead
-- Removed `GridView::renderSelectionForm()` and corresponding `GridView::$selectionRoute`, instead use the action on
-  the selection menu items
+- Removed `GridView::renderSelectionForm()`
 
 ## 2.4.2 (Dec 17, 2024)